- Fixed the smart.
git-svn-id: http://subversion.assembla.com/svn/hyperion_server@216 c48d7634-943c-464e-a392-e2f9434beeda

+++ src/org/hyperion/rs2/net/ActionSender.java
@@ -219,7 +219,7 @@ public class ActionSender {
 	 */
 	public ActionSender sendUpdateItem(int interfaceId, int slot, Item item) {
 		PacketBuilder bldr = new PacketBuilder(34, Type.VARIABLE_SHORT);
-		bldr.putShort(interfaceId).putSmart((byte) slot);
+		bldr.putShort(interfaceId).putSmart(slot);
 		if(item != null) {
 			bldr.putShort(item.getId() + 1);
 			int count = item.getCount();
